requirements.txt
torch==2.5.1
torchvision==0.20.1
pillow==11.1.0
numpy==2.0.2
scikit-learn==1.6.1
matplotlib==3.10.0